# frontend/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Modular-monolith manufacturing ERP — double-entry ledger, inventory, procure-to-pay, order-to-cash, manufacturing and reporting.",
    "title" : "Manufacturing ERP API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/ledger/fiscal-years/{yearCode}/close-year" : {
      "post" : {
        "operationId" : "closeYear",
        "parameters" : [ {
          "in" : "path",
          "name" : "yearCode",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/YearEndCloseResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ledger-controller" ]
      }
    },
    "/api/ledger/fiscal-years/{yearCode}/periods/{periodNo}/close" : {
      "post" : {
        "operationId" : "closePeriod",
        "parameters" : [ {
          "in" : "path",
          "name" : "yearCode",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "periodNo",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FiscalPeriodResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ledger-controller" ]
      }
    },
    "/api/ledger/fiscal-years/{yearCode}/periods/{periodNo}/reopen" : {
      "post" : {
        "operationId" : "reopenPeriod",
        "parameters" : [ {
          "in" : "path",
          "name" : "yearCode",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "periodNo",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FiscalPeriodResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ledger-controller" ]
      }
    },
    "/api/ledger/journal-entries" : {
      "post" : {
        "operationId" : "post",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/JournalEntryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JournalEntryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ledger-controller" ]
      }
    },
    "/api/ledger/journal-entries/{entryNo}" : {
      "get" : {
        "operationId" : "journalEntry",
        "parameters" : [ {
          "in" : "path",
          "name" : "entryNo",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JournalEntryDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ledger-controller" ]
      }
    },
    "/api/ledger/journal-entries/{entryNo}/reverse" : {
      "post" : {
        "operationId" : "reverse",
        "parameters" : [ {
          "in" : "path",
          "name" : "entryNo",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReverseEntryRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JournalEntryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "ledger-controller" ]
      }
    },
...
    "/api/reporting/trial-balance" : {
      "get" : {
        "operationId" : "trialBalance",
        "parameters" : [ {
          "in" : "query",
          "name" : "asOf",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrialBalance"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "reporting-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "FiscalPeriodResponse" : {
        "type" : "object",
        "properties" : {
          "endDate" : {
            "type" : "string",
            "format" : "date"
          },
          "periodNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string"
          },
          "yearCode" : {
            "type" : "string"
          }
        }
      },
...
      "JournalEntryDetail" : {
        "type" : "object",
        "properties" : {
          "currencyCode" : {
            "type" : "string"
          },
          "entryNo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/JournalEntryDetailLine"
            }
          },
          "memo" : {
            "type" : "string"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "reversedByEntryNo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "reversesEntryNo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sourceDocId" : {
            "type" : "string"
          },
          "sourceDocType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "totalCredit" : {
            "type" : "string"
          },
          "totalDebit" : {
            "type" : "string"
          }
        }
      },
      "JournalEntryDetailLine" : {
        "type" : "object",
        "properties" : {
          "accountCode" : {
            "type" : "string"
          },
          "accountName" : {
            "type" : "string"
          },
          "credit" : {
            "type" : "string"
          },
          "debit" : {
            "type" : "string"
          },
          "memo" : {
            "type" : "string"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "JournalEntryLine" : {
        "type" : "object",
        "properties" : {
          "accountCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "credit" : {
            "type" : "string"
          },
          "debit" : {
            "type" : "string"
          },
          "memo" : {
            "type" : "string"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "accountCode", "credit", "debit" ]
      },
      "JournalEntryRequest" : {
        "type" : "object",
        "properties" : {
          "currencyCode" : {
            "type" : "string"
          },
          "journalCode" : {
            "type" : "string"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/JournalEntryLine"
            },
            "minItems" : 1
          },
          "memo" : {
            "type" : "string"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "sourceDocId" : {
            "type" : "string"
          },
          "sourceDocType" : {
            "type" : "string"
          },
          "sourceEvent" : {
            "type" : "string"
          }
        },
        "required" : [ "lines", "postingDate" ]
      },
      "JournalEntryResponse" : {
        "type" : "object",
        "properties" : {
          "currencyCode" : {
            "type" : "string"
          },
          "entryNo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string"
          },
          "totalCredit" : {
            "type" : "string"
          },
          "totalDebit" : {
            "type" : "string"
          }
        }
      },
...
      "ReverseEntryRequest" : {
        "type" : "object",
        "properties" : {
          "memo" : {
            "type" : "string"
          },
          "reversalDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
...
      "TrialBalance" : {
        "type" : "object",
        "properties" : {
          "asOf" : {
            "type" : "string",
            "format" : "date"
          },
          "balanced" : {
            "type" : "boolean"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TrialBalanceRow"
            }
          },
          "totalCredit" : {
            "type" : "string"
          },
          "totalDebit" : {
            "type" : "string"
          }
        }
      },
...
      "TrialBalanceRow" : {
        "type" : "object",
        "properties" : {
          "accountClass" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "credit" : {
            "type" : "string"
          },
          "debit" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "YearEndCloseResult" : {
        "type" : "object",
        "properties" : {
          "closingEntryNo" : {
            "type" : "integer",
            "format" : "int64"
          },
          "netIncome" : {
            "type" : "string"
          },
          "periodsLocked" : {
            "type" : "integer",
            "format" : "int32"
          },
          "yearCode" : {
            "type" : "string"
          },
          "yearStatus" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}